Add authentication scope and namespace to router file
Description:
Added a new scope and namespace for authentication in the router file.

The /auth scope in DiscussWeb routes requests related to authentication. The scope is configured to use the browser pipeline.

Two routes have been defined within the scope:

/:provider for initiating the authentication request
/:provider/callback for handling the callback from the authentication provider
These routes are directed to the AuthController with corresponding actions: request and callback.